+++ src/plugins/stake-plugins/thorchainSavers/tcSaversPlugin.ts
@@ -298,7 +298,14 @@ export const makeTcSaversPlugin = async (opts: EdgeGuiPluginOptions): Promise<St
         throw new Error('pluginId mismatch between request and policy')
       }
 
-      return await changeQuoteFuncs[action](opts, request)
+      switch (action) {
+        case 'stake':
+          return await stakeRequest(opts, request)
+        case 'unstake':
+        case 'claim':
+        case 'unstakeExact':
+          return await unstakeRequest(opts, request)
+      }
     },
     async fetchStakePosition(request: StakePositionRequest): Promise<StakePosition> {
       await updateInboundAddresses(opts)
@@ -583,13 +590,18 @@ const stakeRequest = async (opts: EdgeGuiPluginOptions, request: ChangeQuoteRequ
 
   if (!isToken && lt(addressBalance, nativeAmount)) {
     // Easy check to see if primary address doesn't have enough funds
+    if (isEvm) {
+      // EVM chains only have one address, so if there aren't enough funds in
+      // the primary address then we don't have enough funds at all
+      throw new InsufficientFundsError({ tokenId: null })
+    }
     needsFundingPrimary = true
   } else {
     try {
       const estimateTx = await wallet.makeSpend(spendInfo)
       networkFee = estimateTx.parentNetworkFee ?? estimateTx.networkFee
     } catch (e: unknown) {
-      if (asMaybeInsufficientFundsError(e) != null && !isToken) {
+      if (!isEvm && asMaybeInsufficientFundsError(e) != null) {
         needsFundingPrimary = true
       } else {
         throw e
@@ -877,7 +889,9 @@ const unstakeRequestInner = async (opts: EdgeGuiPluginOptions, request: ChangeQu
   }
 
   const spendInfo: EdgeSpendInfo = {
-    tokenId,
+    // For unstaking we always send just the mainnet coin since we are only sending a message
+    // to the Thorchain pool to withdraw the funds
+    tokenId: null,
     spendTargets: [{ publicAddress: poolAddress, nativeAmount: sendNativeAmount }],
     otherParams: { enableRbf: false, outputSort: 'targets', utxoSourceAddress, forceChangeAddress },
     assetAction: { assetActionType: 'unstakeOrder' },
@@ -912,6 +926,12 @@ const unstakeRequestInner = async (opts: EdgeGuiPluginOptions, request: ChangeQu
 
   if (lt(balanceToCheck, sendNativeAmount)) {
     // Easy check to see if primary address doesn't have enough funds
+    if (isEvm) {
+      // EVM chains only have one address, so if there aren't enough funds in
+      // the primary address then we don't have enough funds at all
+      throw new InsufficientFundsError({ tokenId: null })
+    }
+    // Easy check to see if primary address doesn't have enough funds
     needsFundingPrimary = true
   } else {
     try {
@@ -919,7 +939,7 @@ const unstakeRequestInner = async (opts: EdgeGuiPluginOptions, request: ChangeQu
       const estimateTx = await wallet.makeSpend(spendInfo)
       networkFee = estimateTx.networkFee
     } catch (e: unknown) {
-      if (asMaybeInsufficientFundsError(e) != null) {
+      if (!isEvm && asMaybeInsufficientFundsError(e) != null) {
         needsFundingPrimary = true
       } else {
         throw e
@@ -1020,13 +1040,6 @@ const unstakeRequestInner = async (opts: EdgeGuiPluginOptions, request: ChangeQu
       await wallet.saveTx(broadcastedTx)
     }
   }
-}
-
-const changeQuoteFuncs = {
-  stake: stakeRequest,
-  unstake: unstakeRequest,
-  claim: unstakeRequest,
-  unstakeExact: unstakeRequest
 }
 
 const headers = {